# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpdQueueCreateRequest": {
        "properties": {
          "opd_visit_id": {
            "format": "uuid",
            "title": "Opd Visit Id",
            "type": "string"
          },
          "priority": {
            "default": "normal",
            "enum": [
              "normal",
              "urgent"
            ],
            "title": "Priority",
            "type": "string"
          }
        },
        "required": [
          "opd_visit_id"
        ],
        "title": "OpdQueueCreateRequest",
        "type": "object"
      },
      "OpdQueueSkipRequest": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OpdQueueSkipRequest",
        "type": "object"
      },
      "OpdQueueUpdateRequest": {
        "properties": {
          "position": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "priority": {
            "anyOf": [
              {
                "enum": [
                  "normal",
                  "urgent"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          }
        },
        "title": "OpdQueueUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Management SaaS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opd/queue": {
      "get": {
        "description": "Poll doctor queue for a date.",
        "operationId": "get_queue_board_api_v1_opd_queue_get",
        "parameters": [
          {
            "description": "Doctor whose queue to fetch",
            "in": "query",
            "name": "doctor_id",
            "required": true,
            "schema": {
              "description": "Doctor whose queue to fetch",
              "format": "uuid",
              "title": "Doctor Id",
              "type": "string"
            }
          },
          {
            "description": "Queue date; defaults to today",
            "in": "query",
            "name": "queue_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Queue date; defaults to today",
              "title": "Queue Date"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Queue Board Api V1 Opd Queue Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Queue Board",
        "tags": [
          "OPD"
        ]
      },
      "post": {
        "description": "Add visit to doctor queue.",
        "operationId": "add_to_queue_api_v1_opd_queue_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdQueueCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Add To Queue Api V1 Opd Queue Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add To Queue",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/queue/{queue_id}": {
      "patch": {
        "description": "Reorder queue or change priority.",
        "operationId": "update_queue_entry_api_v1_opd_queue__queue_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "queue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queue Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdQueueUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Queue Entry Api V1 Opd Queue  Queue Id  Patch",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Queue Entry",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/queue/{queue_id}/call": {
      "post": {
        "description": "Call next patient in queue.",
        "operationId": "call_queue_patient_api_v1_opd_queue__queue_id__call_post",
        "parameters": [
          {
            "in": "path",
            "name": "queue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queue Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Call Queue Patient Api V1 Opd Queue  Queue Id  Call Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Call Queue Patient",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/queue/{queue_id}/complete": {
      "post": {
        "description": "Mark queue entry complete.",
        "operationId": "complete_queue_entry_api_v1_opd_queue__queue_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "queue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queue Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Complete Queue Entry Api V1 Opd Queue  Queue Id  Complete Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Queue Entry",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/queue/{queue_id}/skip": {
      "post": {
        "description": "Skip patient in queue.",
        "operationId": "skip_queue_patient_api_v1_opd_queue__queue_id__skip_post",
        "parameters": [
          {
            "in": "path",
            "name": "queue_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Queue Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdQueueSkipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Skip Queue Patient Api V1 Opd Queue  Queue Id  Skip Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Skip Queue Patient",
        "tags": [
          "OPD"
        ]
      }
    },
...
  }
}